Add automatic socket name detection in compositor
The compositor automatically finds a free socket name using
wl_display_add_socket_auto. Also fix the order in which the compositor
uses --wayland-socket-name or the content of the socketName property.
The documentation states the content of the socketName property should
be used before the content of --wayland-socket-name.

The commit also removes the reference to the WAYLAND_DISPLAY
environment variable from the docs.

Change-Id: I33dff4e4627d95f5894a5a1e2f308d747f449691